General improvement in the organization of the Code - Separate UI logic from Robot control - Add new class SimulationEnviroment, to handle coins , collisions, render, etc... - Update demo - Folder name change to be more cohesive
