+++ demo/src/index.ts
@@ -2,103 +2,11 @@ import { buildHex } from './compile';
 import './index.css';
 import { EditorHistoryUtil } from './utils/editor-history.util';
 import {Robots} from "../../src";
+import {MOVE_CODE} from "./sketch";
+
 console.log("hello")
 let editor: any; // eslint-disable-line @typescript-eslint/no-explicit-any
-const MOVE_CODE = `#include <Servo.h>
-Servo leftservo;  
-Servo rightservo;  
 
-void setup() {
-  Serial.begin(9600);
-  leftservo.attach(9);  
-  rightservo.attach(10); 
-  //move forward fast
-  leftservo.write(90);
-  rightservo.write(90);
-
-  delay(500);
-  leftservo.write(180);
-  rightservo.write(0);
-  delay(1000);
- 
-}
-void loop() {
-  float leftDistance = getDistance(2,3);
-  float frontDistance = getDistance(4,5);
-  Serial.println(leftDistance);
-  Serial.println(frontDistance);
-  if(frontDistance < 50)
-  {
-      stop();
-      return;
-  }
-  if(leftDistance < 180)
-    moveAwayFromTheLeftWall();
-  else if(leftDistance > 220)
-    moveCloserToTheLeftWall();
-  else
-    moveForward();
-
-  delay(10);
-}
-void stop(){
-  leftservo.write(90);
-  rightservo.write(90);
-}
-void moveForward(){
-  leftservo.write(170);
-  rightservo.write(10);
-  
-}
-void moveAwayFromTheLeftWall(){
-  leftservo.write(160);
-  rightservo.write(160);
-
-  delay(600);
-
- leftservo.write(160);
-  rightservo.write(20);
-  delay(500);
-
- leftservo.write(20);
-  rightservo.write(20);
-
-  delay(600);
-  leftservo.write(90);
-  rightservo.write(90);
-}
-
-void moveCloserToTheLeftWall(){
-    leftservo.write(20);
-  rightservo.write(20);
-
-  delay(600);
-
- leftservo.write(160);
-  rightservo.write(20);
-  delay(500);
-
- leftservo.write(160);
-  rightservo.write(160);
-
-  delay(600);
-  leftservo.write(90);
-  rightservo.write(90);
-}
-
-int getDistance(int trigger, int echo){
-  long duration;  
-  //clear the ping pin
-  digitalWrite(trigger, LOW);
-  delayMicroseconds(2);
-  //send the 10 microsecond trigger
-  digitalWrite(trigger, HIGH);
-  delayMicroseconds(10);
-  digitalWrite(trigger, LOW);
-  //get the pulse duration in microseconds
-  duration = pulseIn(echo, HIGH);
-  return duration/ 29 / 2;
-}`;
 
 // Load Editor
 declare const window: any; // eslint-disable-line @typescript-eslint/no-explicit-any
@@ -135,25 +43,29 @@ function initiateRobot(){
 //set up robot environment
     const canvas = document.getElementById('world');
 
-    const robot = new Robots.Arduino.TwoServoRobot(canvas)
-    robot.addUltrasonicSensor('CFL', 2,3);
+    const robot = new Robots.Arduino.TwoServoRobot(9, 10);
+    const env =  new  Robots.SimulationEnviroment(robot, canvas);
+
+    robot.addUltrasonicSensor('CFL', 2,3); // triggerPin,echoPin
     robot.addUltrasonicSensor('F2', 4,5);
-    robot.addUltrasonicSensor('CFR', 2,3);
-    robot.addUltrasonicSensor('CBR', 4,5);robot.addUltrasonicSensor('CFL', 2,3);
-    robot.addUltrasonicSensor('CBL', 4,5);robot.addUltrasonicSensor('CFL', 2,3);
+    robot.addUltrasonicSensor('CFR', 6,7);
+    //robot.addUltrasonicSensor('CBR', 4,5);
+    //robot.addUltrasonicSensor('CBL', 4,5);
     robot.environment.setRobotInitialPosition({x:50, y: 100});
-    robot.environment?.addObstacleRectangle(400, 50, 800, 20, "grey");
-    robot.environment?.addObstacleRectangle(700, 200, 20, 800, "grey");
+    
+    env.addObstacleRectangle(400, 50, 800, 20, "grey");
+    env.addObstacleRectangle(700, 200, 20, 800, "grey");
 
-    robot.environment?.addCoin(150,100);
-    robot.environment?.addCoin(300, 100);
+    env.addCoin(150,100);
+    env.addCoin(200, 150);
 
-    robot.environment.OnAllCoinsCollectedEvent = (env) => {
+    env.OnAllCoinsCollectedEvent = (env) => {
         env.removedCoins = [];
         env.coins = [];
         env.addCoin(400, 200);
         env.addCoin(500,500);
     }
+    
     robot.arduino.simulationTimeCallback = (time)=>{
         statusLabel.textContent = `Simulation time: ${time}`;
     }
