+++ src/enviroment/utils/utils.ts
