--- src/robots/index.ts
@@ -1,1 +1,0 @@
-export {TwoWheelRobot} from './TwoWheelRobot';
